refactor: CatalogItem 모델 위치 변경
