+++ app/src/main/java/woowacourse/shopping/presentation/view/catalog/CatalogFragment.kt
@@ -11,10 +11,10 @@ import woowacourse.shopping.R
 import woowacourse.shopping.databinding.FragmentCatalogBinding
 import woowacourse.shopping.databinding.MenuItemCartBinding
 import woowacourse.shopping.presentation.base.BaseFragment
+import woowacourse.shopping.presentation.model.CatalogItem
 import woowacourse.shopping.presentation.ui.decorations.GridSpacingItemDecoration
 import woowacourse.shopping.presentation.view.cart.CartFragment
 import woowacourse.shopping.presentation.view.catalog.adapter.CatalogAdapter
-import woowacourse.shopping.presentation.view.catalog.adapter.model.CatalogItem
 import woowacourse.shopping.presentation.view.catalog.event.CatalogMessageEvent
 import woowacourse.shopping.presentation.view.detail.DetailFragment
